+++ organization/src/epilot/sdk.py
@@ -7,7 +7,7 @@ from .organization_settings import OrganizationSettings
 from .sdkconfiguration import SDKConfiguration
 from epilot import utils
 from epilot.models import components
-from typing import Dict
+from typing import Callable, Dict, Union
 
 class Epilot:
     r"""Organization API: Manage epilot tenant organizations"""
@@ -20,7 +20,7 @@ class Epilot:
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 epilot_auth: str ,
+                 epilot_auth: Union[str, Callable[[], str]],
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,
@@ -30,7 +30,7 @@ class Epilot:
         """Instantiates the SDK configuring it with the provided parameters.
         
         :param epilot_auth: The epilot_auth required for authentication
-        :type epilot_auth: Union[str,Callable[[], str]]
+        :type epilot_auth: Union[str, Callable[[], str]]
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations
@@ -45,7 +45,11 @@ class Epilot:
         if client is None:
             client = requests_http.Session()
         
-        security = components.Security(epilot_auth = epilot_auth)
+        if callable(epilot_auth):
+            def security():
+                return components.Security(epilot_auth = epilot_auth())
+        else:
+            security = components.Security(epilot_auth = epilot_auth)
         
         if server_url is not None:
             if url_params is not None: