+++ organization/src/epilot/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.4.3'
-    gen_version: str = '2.233.2'
-    user_agent: str = 'speakeasy-sdk/python 1.4.3 2.233.2 1.0.0 epilot-organization'
+    sdk_version: str = '1.4.4'
+    gen_version: str = '2.234.3'
+    user_agent: str = 'speakeasy-sdk/python 1.4.4 2.234.3 1.0.0 epilot-organization'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: